Revert "Feat/types новая версия скрипта и добавление новых типов, под новое api"
