+++ package.json
@@ -8,7 +8,8 @@
     "build": "tsc -b && vite build",
     "lint": "eslint .",
     "preview": "vite preview",
-    "build-gen:all": "tsc -P tsconfig.types.json && node src/entities/gen-types.js && eslint -c eslint.config.js src/entities --fix && prettier --write src/entities",
+    "compile:splitjs": "tsc -P tsconfig.types.json",
+    "gen:types": "node src/entities/ts-split-types.js",
     "build:docker": "docker build -t caselab2024-team2 .",
     "start:docker": "docker run -d -p 5173:80 --rm --name caselab2024-team2 caselab2024-team2",
     "stop:docker": "docker stop caselab2024-team2",
