+++ src/entities/Attribute/index.ts
@@ -1,4 +1,2 @@
-export type * from './model/types/AttributeValueRequest';
-export type * from './model/types/AttributeValueResponse';
-export type * from './model/types/AttributeRequest';
-export type * from './model/types/AttributeResponse';
+export * from './model/AttributeRequest';
+export * from './model/AttributeResponse';
