+++ src/entities/Attribute/model/AttributeRequest.ts
@@ -1,0 +1,4 @@
+export interface AttributeRequest {
+name: string; // Имя атрибута
+type: string; // Тип атрибута
+}
