+++ src/entities/Attribute/model/AttributeResponse.ts
@@ -1,0 +1,5 @@
+export interface AttributeResponse {
+id: number; // ID атрибута
+name: string; // Имя атрибута
+type: string; // Тип атрибута
+}
