--- src/entities/Attribute/model/types/AttributeRequest.ts
@@ -1,4 +1,0 @@
-export interface AttributeRequest {
-  name: string;
-  type: string;
-}
