--- src/entities/Attribute/model/types/AttributeValueRequest.ts
@@ -1,4 +1,0 @@
-export interface AttributeValueRequest {
-  attributeId: number;
-  value: string;
-}
