--- src/entities/Attribute/model/types/AttributeValueResponse.ts
@@ -1,6 +1,0 @@
-export interface AttributeValueResponse {
-  id: number;
-  name: string;
-  type: string;
-  value: string;
-}
