+++ src/entities/Document/index.ts
@@ -1,11 +1,7 @@
-export type * from './model/types/UpdateDocumentRequest';
-export type * from './model/types/DocumentFacadeResponse';
-export type * from './model/types/DocumentPermissionResponse';
-export type * from './model/types/DocumentResponse';
-export type * from './model/types/DocumentVersionResponse';
-export type * from './model/types/DocumentTypeRequest';
-export type * from './model/types/DocumentTypeToAttributeRequest';
-export type * from './model/types/DocumentTypeResponse';
-export type * from './model/types/DocumentTypeToAttributeResponse';
-export type * from './model/types/CreateDocumentRequest';
-export type * from './model/types/PatchDocumentRequest';
+export * from './model/DocumentRequest';
+export * from './model/DocumentPermissionResponse';
+export * from './model/DocumentResponse';
+export * from './model/DocumentTypeRequest';
+export * from './model/DocumentTypeToAttributeRequest';
+export * from './model/DocumentTypeResponse';
+export * from './model/DocumentTypeToAttributeResponse';
