+++ src/entities/Document/model/DocumentPermissionResponse.ts
@@ -1,0 +1,4 @@
+export interface DocumentPermissionResponse {
+id: number; // ID разрешения
+name: string; // Название разрешения
+}
