+++ src/entities/Document/model/DocumentRequest.ts
@@ -1,0 +1,6 @@
+import { UserToDocumentRequest } from '../../User/model/UserToDocumentRequest.ts';
+export interface DocumentRequest {
+document_type_id: number; // id типа документа
+name: string; // Имя документа
+users_permissions: UserToDocumentRequest[]; // 
+}
