+++ src/entities/Document/model/DocumentResponse.ts
@@ -1,0 +1,8 @@
+import { UserToDocumentResponse } from '../../User/model/UserToDocumentResponse.ts';
+export interface DocumentResponse {
+id: number; // ID документа
+document_type_id: number; // ID типа документа
+name: string; // Имя документа
+document_versions_ids: number[]; // 
+user_permissions: UserToDocumentResponse[]; // 
+}
