+++ src/entities/Document/model/DocumentTypeRequest.ts
@@ -1,0 +1,5 @@
+import { DocumentTypeToAttributeRequest } from '../../Document/model/DocumentTypeToAttributeRequest.ts';
+export interface DocumentTypeRequest {
+name: string; // Название типа документа
+attributes: DocumentTypeToAttributeRequest[]; // 
+}
