+++ src/entities/Document/model/DocumentTypeResponse.ts
@@ -1,0 +1,6 @@
+import { DocumentTypeToAttributeResponse } from '../../Document/model/DocumentTypeToAttributeResponse.ts';
+export interface DocumentTypeResponse {
+attributes: DocumentTypeToAttributeResponse[]; // 
+id: number; // ID типа документа
+name: string; // Название типа документа
+}
