+++ src/entities/Document/model/DocumentTypeToAttributeRequest.ts
@@ -1,0 +1,4 @@
+export interface DocumentTypeToAttributeRequest {
+attribute_id: number; // ID аттрибута
+is_optional: boolean; // Признак обязательности аттрибута
+}
