+++ src/entities/Document/model/DocumentTypeToAttributeResponse.ts
@@ -1,0 +1,4 @@
+export interface DocumentTypeToAttributeResponse {
+attribute_id: number; // ID аттрибута
+is_optional: boolean; // Признак обязательности аттрибута
+}
