--- src/entities/Document/model/types/DocumentPermissionResponse.ts
@@ -1,4 +1,0 @@
-export interface DocumentPermissionResponse {
-  id: number;
-  name: string;
-}
