--- src/entities/Document/model/types/DocumentTypeRequest.ts
@@ -1,5 +1,0 @@
-import type { DocumentTypeToAttributeRequest } from '../../../Document/model/types/DocumentTypeToAttributeRequest.ts';
-export interface DocumentTypeRequest {
-  name: string;
-  attributes: DocumentTypeToAttributeRequest;
-}
