--- src/entities/Document/model/types/DocumentTypeResponse.ts
@@ -1,6 +1,0 @@
-import type { DocumentTypeToAttributeResponse } from '../../../Document/model/types/DocumentTypeToAttributeResponse.ts';
-export interface DocumentTypeResponse {
-  attributes: DocumentTypeToAttributeResponse;
-  id: number;
-  name: string;
-}
