--- src/entities/Document/model/types/DocumentTypeToAttributeRequest.ts
@@ -1,4 +1,0 @@
-export interface DocumentTypeToAttributeRequest {
-  attribute_id: number;
-  is_optional: boolean;
-}
