--- src/entities/Document/model/types/DocumentTypeToAttributeResponse.ts
@@ -1,4 +1,0 @@
-export interface DocumentTypeToAttributeResponse {
-  attribute_id: number;
-  is_optional: boolean;
-}
