--- src/entities/Document/model/types/DocumentVersionResponse.ts
@@ -1,11 +1,0 @@
-import type { AttributeValueResponse } from '../../../Attribute/model/types/AttributeValueResponse.ts';
-export interface DocumentVersionResponse {
-  id: number;
-  attributes: AttributeValueResponse;
-  name: string;
-  createdAt: string;
-  documentId: number;
-  signatureIds: number[];
-  votingProcessesId: number[];
-  contentName: string;
-}
