--- src/entities/Document/model/types/PatchDocumentRequest.ts
@@ -1,6 +1,0 @@
-import type { AttributeValueRequest } from '../../../Attribute/model/types/AttributeValueRequest.ts';
-export interface PatchDocumentRequest {
-  document_type_id: number;
-  name: string;
-  version_attributes: AttributeValueRequest;
-}
