+++ src/entities/Signature/model/SignatureCreateRequest.ts
@@ -1,0 +1,5 @@
+export interface SignatureCreateRequest {
+documentVersionId: number; // ID версии документа
+name: string; // Название подписи
+email: string; // Адрес электронной почты пользователя
+}
