--- src/entities/Signature/model/types/SignatureCreateRequest.ts
@@ -1,5 +1,0 @@
-export interface SignatureCreateRequest {
-  documentId: number;
-  name: string;
-  email: string;
-}
