--- src/entities/Signature/model/types/SignatureResponse.ts
@@ -1,10 +1,0 @@
-export interface SignatureResponse {
-  id: number;
-  name: string;
-  status: string;
-  sentAt: string;
-  signedAt: string;
-  signatureData: string;
-  email: string;
-  documentId: number;
-}
