+++ src/entities/User/index.ts
@@ -1,6 +1,7 @@
-export type * from './model/types/UserUpdateRequest';
-export type * from './model/types/UserResponse';
-export type * from './model/types/UserToDocumentResponse';
-export type * from './model/types/RegisterRequest';
-export type * from './model/types/AuthenticationRequest';
-export type * from './model/types/AuthenticationResponse';
+export * from './model/UserUpdateRequest';
+export * from './model/UserResponse';
+export * from './model/UserToDocumentRequest';
+export * from './model/UserToDocumentResponse';
+export * from './model/RegisterRequest';
+export * from './model/AuthenticationResponse';
+export * from './model/AuthenticationRequest';
