+++ src/entities/User/model/AuthenticationRequest.ts
@@ -1,0 +1,4 @@
+export interface AuthenticationRequest {
+email: string; // Адрес электронной почты пользователя
+password: string; // Пароль пользователя
+}
