+++ src/entities/User/model/AuthenticationResponse.ts
@@ -1,0 +1,3 @@
+export interface AuthenticationResponse {
+token: string; // JWT токен для аутентифицированного пользователя
+}
