+++ src/entities/User/model/RegisterRequest.ts
@@ -1,0 +1,5 @@
+export interface RegisterRequest {
+display_name: string; // Отображаемое имя пользователя
+email: string; // Адрес электронной почты пользователя
+password: string; // Пароль пользователя
+}
