+++ src/entities/User/model/UserToDocumentRequest.ts
@@ -1,0 +1,4 @@
+export interface UserToDocumentRequest {
+email: string; // Email пользователя
+document_permissions: number[]; // 
+}
