+++ src/entities/User/model/UserToDocumentResponse.ts
@@ -1,0 +1,5 @@
+import { DocumentPermissionResponse } from '../../Document/model/DocumentPermissionResponse.ts';
+export interface UserToDocumentResponse {
+email: string; // Email пользователя
+document_permissions: DocumentPermissionResponse[]; // 
+}
