+++ src/entities/User/model/UserUpdateRequest.ts
@@ -1,0 +1,4 @@
+export interface UserUpdateRequest {
+display_name: string; // Отображаемое имя пользователя
+password: string; // Пароль пользователя
+}
