--- src/entities/User/model/types/AuthenticationRequest.ts
@@ -1,4 +1,0 @@
-export interface AuthenticationRequest {
-  email: string;
-  password: string;
-}
