--- src/entities/User/model/types/AuthenticationResponse.ts
@@ -1,3 +1,0 @@
-export interface AuthenticationResponse {
-  token: string;
-}
