--- src/entities/User/model/types/UserResponse.ts
@@ -1,5 +1,0 @@
-export interface UserResponse {
-  email: string;
-  display_name: string;
-  roles: string[];
-}
