--- src/entities/User/model/types/UserToDocumentResponse.ts
@@ -1,5 +1,0 @@
-import type { DocumentPermissionResponse } from '../../../Document/model/types/DocumentPermissionResponse.ts';
-export interface UserToDocumentResponse {
-  email: string;
-  document_permissions: DocumentPermissionResponse;
-}
