--- src/entities/Vote/index.ts
@@ -1,3 +1,0 @@
-export type * from './model/types/VoteResponse';
-export type * from './model/types/VoteUserResponse';
-export type * from './model/types/VoteRequest';
