--- src/entities/Vote/model/types/VoteResponse.ts
@@ -1,5 +1,0 @@
-import type { VoteUserResponse } from '../../../Vote/model/types/VoteUserResponse.ts';
-export interface VoteResponse {
-  status: string;
-  applicationUser: VoteUserResponse;
-}
