--- src/entities/Vote/model/types/VoteUserResponse.ts
@@ -1,4 +1,0 @@
-export interface VoteUserResponse {
-  email: string;
-  displayName: string;
-}
