--- src/entities/Voting/index.ts
@@ -1,2 +1,0 @@
-export type * from './model/types/VotingProcessRequest';
-export type * from './model/types/VotingProcessResponse';
