--- src/entities/Voting/model/types/VotingProcessRequest.ts
@@ -1,7 +1,0 @@
-export interface VotingProcessRequest {
-  name: string;
-  threshold: number;
-  deadline: { [key: string]: unknown };
-  documentId: number;
-  emails: string[];
-}
