+++ src/shared/types/ProblemDetail.ts
@@ -1,8 +1,8 @@
 export interface ProblemDetail {
-	type: string;
-	title: string;
-	status: number;
-	detail: string;
-	instance: string;
-	properties: { [key: string]: unknown };
+type: string; // 
+title: string; // 
+status: number; // 
+detail: string; // 
+instance: string; // 
+properties: { [key: string]: any }; // 
 }
